requirements.txt
asgiref==3.5.0
awesomeversion==21.2.3
billiard==3.6.4.0
cryptography==3.3.2
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.7.0
django-extensions==3.0.8
django-js-asset==1.2.2
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
PyYAML==5.4.1
whitenoise==6.0.0
django-rest-auth[with_social]
djangorestframework
django-rest-knox
youtube-search-python
django-embed-video
pafy
youtube_dl==2020.12.2
tqdm==4.36.1
celery==5.1.2
mutagen==1.45.1
django-picklefield==3.0.1
python-magic==0.4.18; sys_platform == 'linux'
python-magic-bin==0.4.14; sys_platform != 'linux'
yt-dlp@git+https://github.com/JeffreyCA/yt-dlp@a31e44e365e4c5ece2b04df3f1dc146076889fd2
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
youtube-title-parse==1.0.0
websockets==10.0
gevent
requests==2.27.1
redis==3.5.3
kombu==5.2.3
ffmpeg-python==0.2.0
youtube-dl
spleeter
tensorflow-cpu
librosa==0.8.0
llvmlite==0.36.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
dropbox==9.0.0
